backend/requirements.txt
asgi_lifespan
asyncio
black
colorama
email-validator
fastapi
greenlet
h5py
httpx
imageio
isort
loguru
mypy
pandas
pillow
pre-commit
python-dotenv
python-decouple
python-jose
python-multipart
python-slugify
pytest-asyncio
pytest-cov
pytest-xdist
scikit-learn
tensorflow
trio
uvicorn
